+++ Code/Engine/Foundation/Configuration/Implementation/CVar.cpp
@@ -399,7 +399,7 @@ void ezCVar::LoadCVarsFromFile(bool bOnlyNewOnes, bool bSetAsCurrentValue)
                 {
                   ezCVarFloat* pTyped = (ezCVarFloat*)pCVar;
                   pTyped->m_Values[ezCVarValue::Stored] = static_cast<float>(Value);
-                  *pTyped = Value;
+                  *pTyped = static_cast<float>(Value);
                 }
               }
               break;
@@ -474,7 +474,7 @@ void ezCVar::LoadCVarsFromCommandLine(bool bOnlyNewOnes /*= true*/, bool bSetAsC
           Value = ezCommandLineUtils::GetGlobalInstance()->GetFloatOption(sTemp, Value);
 
           pTyped->m_Values[ezCVarValue::Stored] = static_cast<float>(Value);
-          *pTyped = Value;
+          *pTyped = static_cast<float>(Value);
         }
         break;
         case ezCVarType::String:
